+++ CHANGELOG.md
@@ -41,6 +41,7 @@ and this project adheres to
 - ⬆️(dependencies) update vite to v7.3.2 [SECURITY]
 - ⬆️(dependencies) update django to v5.2.13 [SECURITY]
 - 🔒(backend) rely on backend to allow participant update their metadata
+- 🩹(backend) fix swagger and redoc documentation URLs
 - 🐛(summary) fix failure webhook notification #1233
 - 🐛(summary) relax whisperX payload format #1233
 - ⬆️(backend) upgrade dependencies to fix Pillow CVE-2026-40192